added validation for users

+++ controllers/userControllers.js
@@ -3,6 +3,8 @@ const User = require('../models/user')
 exports.allUserDetails = async (req, res) => {
     try {
         const users = await User.find().select('-cart -__v -_id');
+        if (!users)
+            return res.status(404).json({ message: "Something unexpected is requested" })
         res.status(201).json(users);
     } catch (err) {
         res.status(500).json({ message: err.message });
@@ -14,7 +16,7 @@ exports.createUser = async (req, res) => {
     try {
         const user = new User(req.body);
         const result = await user.save();
-        res.status(201).json({message:"New user created successfully"});
+        res.status(201).json({ message: "New user created successfully" });
     } catch (err) {
         res.status(500).json({ message: err.message });
     }
@@ -23,6 +25,8 @@ exports.createUser = async (req, res) => {
 exports.userDetailById = async (req, res) => {
     try {
         const user = await User.findById(req.params.id).select('-cart -__v -_id');
+        if (!user)
+            return res.status(404).json({ message: "Requested user doesn't exist" });
         res.status(201).json(user);
     } catch (err) {
         res.status(500).json({ message: err.message });
@@ -32,6 +36,9 @@ exports.userDetailById = async (req, res) => {
 
 exports.deleteUserById = async (req, res) => {
     try {
+        const userId = await User.findById(req.params.id);
+        if (!userId)
+            return res.status(404).json({ message: "Requested user doesn't exist" });
         const user = await User.findByIdAndDelete(req.params.id);
         res.status(201).json(user);
     } catch (err) {
@@ -41,8 +48,13 @@ exports.deleteUserById = async (req, res) => {
 
 exports.updateUser = async (req, res) => {
     try {
-        const { username, email, phoneNo } = req.body
-        const user = await User.findByIdAndUpdate({_id:req.params.id}, {username:username, email:email,phoneNo:phoneNo}, {new:true});
+        const { username, email, phoneNo } = req.body;
+        const userId = await User.findById(req.params.id);
+        
+        if (!userId)
+            return res.status(404).json({ message: "Requested user doesn't exist" });
+
+        const user = await User.findByIdAndUpdate({ _id: req.params.id }, { username: username, email: email, phoneNo: phoneNo }, { new: true });
         res.status(201).json(user);
     } catch (err) {
         res.status(500).json({ message: err.message });
